==> 02_create_word_plots.R <==
library(here)
library(tidyverse)
source(here::here("00_helpers.R"))
library(ggsignif)


## overall 
plot_word_by_word("Montevideo", "BFD1")
plot_word_by_word("Durazno", "BFD1")

plot_word_by_word("Montevideo", "BFD2")
plot_word_by_word("Durazno", "BFD2")

plot_word_by_word("Montevideo", "BFD4")
plot_word_by_word("Durazno", "BFD4")

plot_word_by_word("Montevideo", "NFD1")
plot_word_by_word("Durazno", "NFD1")

plot_word_by_word("Montevideo", "NFD2")
plot_word_by_word("Durazno", "NFD2")

plot_word_by_word("Montevideo", "NFD3")
plot_word_by_word("Durazno", "NFD3")

plot_word_by_word("Montevideo", "NFD4")
plot_word_by_word("Durazno", "NFD4")

plot_word_by_word("Montevideo", "NFD5")
plot_word_by_word("Durazno", "NFD5")

plot_word_by_word("Montevideo", "NFD6")
plot_word_by_word("Durazno", "NFD6")

# plots by sex 

plot_word_by_word_sex("Montevideo", "BFD1")
plot_word_by_word_sex("Durazno", "BFD1")

plot_word_by_word_sex("Montevideo", "BFD2")
plot_word_by_word_sex("Durazno", "BFD2")

plot_word_by_word_sex("Montevideo", "BFD4")
plot_word_by_word_sex("Durazno", "BFD4")

plot_word_by_word_sex("Montevideo", "NFD1")
plot_word_by_word_sex("Durazno", "NFD1")

plot_word_by_word_sex("Montevideo", "NFD2")
plot_word_by_word_sex("Durazno", "NFD2")

plot_word_by_word_sex("Montevideo", "NFD3")
plot_word_by_word_sex("Durazno", "NFD3")

plot_word_by_word_sex("Montevideo", "NFD4")
plot_word_by_word_sex("Durazno", "NFD4")

plot_word_by_word_sex("Montevideo", "NFD5")
plot_word_by_word_sex("Durazno", "NFD5")

plot_word_by_word_sex("Montevideo", "NFD6")
plot_word_by_word_sex("Durazno", "NFD6")

# Education 

# plots by education

plot_word_by_word_education("Montevideo", "BFD1")
plot_word_by_word_education("Durazno", "BFD1")

plot_word_by_word_education("Montevideo", "BFD2")
plot_word_by_word_education("Durazno", "BFD2")

plot_word_by_word_education("Montevideo", "BFD4")
plot_word_by_word_education("Durazno", "BFD4")

plot_word_by_word_education("Montevideo", "NFD1")
plot_word_by_word_education("Durazno", "NFD1")

plot_word_by_word_education("Montevideo", "NFD2")
plot_word_by_word_education("Durazno", "NFD2")

plot_word_by_word_education("Montevideo", "NFD3")
plot_word_by_word_education("Durazno", "NFD3")

plot_word_by_word_education("Montevideo", "NFD4")
plot_word_by_word_education("Durazno", "NFD4")

plot_word_by_word_education("Montevideo", "NFD5")
plot_word_by_word_education("Durazno", "NFD5")

plot_word_by_word_education("Montevideo", "NFD6")
plot_word_by_word_education("Durazno", "NFD6")



# Age 

# plots by age

plot_word_by_word_age("Montevideo", "BFD1")
plot_word_by_word_age("Durazno", "BFD1")

plot_word_by_word_age("Montevideo", "BFD2")
plot_word_by_word_age("Durazno", "BFD2")

plot_word_by_word_age("Montevideo", "BFD4")
plot_word_by_word_age("Durazno", "BFD4")

plot_word_by_word_age("Montevideo", "NFD1")
plot_word_by_word_age("Durazno", "NFD1")

plot_word_by_word_age("Montevideo", "NFD2")
plot_word_by_word_age("Durazno", "NFD2")

plot_word_by_word_age("Montevideo", "NFD3")
plot_word_by_word_age("Durazno", "NFD3")

plot_word_by_word_age("Montevideo", "NFD4")
plot_word_by_word_age("Durazno", "NFD4")

plot_word_by_word_age("Montevideo", "NFD5")
plot_word_by_word_age("Durazno", "NFD5")

plot_word_by_word_age("Montevideo", "NFD6")
plot_word_by_word_age("Durazno", "NFD6")
